refactor(finance): split finance utils into focused modules
Split the 2,347-line investor_app/finance/utils.py into mortgage,
taxes, scoring, and strategies submodules; utils.py now holds only
core KPI primitives and the Django-coupled analysis. Update all 17
importers to canonical locations.

Resolve LIMIT-20: delete the calculate_* aliases, the dead
score_listing_v1 chain, the duplicate pure score_listing_v2, and the
unused service-layer calculate_noi in property_service.py. Mark
LIMIT-20 resolved in docs/KNOWN_LIMITATIONS.md.

Full suite: 1803 passed. Ruff + mypy clean on touched files.
